[Improvement] Configuration 제거 및 Bezier prefix 추가
